ref: カスタムフック化
